Fix Ctrl-R and Ctrl-S prompt that does not accept space character as part of search string.
The space character generates keyinfo events with keyname defined. Such events are
filtered out by IncrementalSearchPromptMode._process_incremental_search_keyevent() if these are
not considered special. While this is true for most characters (unicode and named national characters
to be tested), space is the special case and it is wanted as part of the search string.